requirements.txt
# CLaC Requirements
# Installation order is important - follow the instructions in readme.md

# ===== Core Deep Learning Framework =====
# PyTorch with CUDA 12.6 support
torch==2.8.0
torchvision==0.23.0

# ===== Graph Neural Networks =====
# PyTorch Geometric and extensions
# Install from PyG repository for proper CUDA compatibility
torch-geometric==2.6.1
pyg-lib
torch-scatter
torch-sparse
torch-cluster
torch-spline_conv

# ===== Materials Science =====
# Orbital graph models and materials database tools
orb-models==0.5.5
jarvis-tools==2025.5.30

# ===== Natural Language Processing =====
# Transformers and datasets for LLM integration
transformers==4.56.2
datasets==4.1.1
spacy==3.8.7

# ===== Training Infrastructure =====
# PyTorch Lightning for training management
lightning==2.5.5

# ===== Configuration and Logging =====
# Hydra for configuration management, W&B for experiment tracking
hydra-core==1.3.2
wandb==0.22.0

# ===== Scientific Computing =====
# Standard scientific Python stack
matplotlib>=3.5.0
scikit-learn>=1.0.0
scipy>=1.7.0
pandas>=1.3.0
numpy>=1.21.0

# ===== Utilities =====
# Additional utilities
python-dotenv>=0.19.0
tqdm>=4.62.0
joblib>=1.1.0
PyYAML>=6.0

# ===== Dimensionality Reduction =====
# For embedding visualization
umap-learn>=0.5.3

# Note: Install PyTorch and PyTorch Geometric with specific CUDA versions
# See readme.md for detailed installation instructions
